fix(cli): keep SessionDelete hook dispatch ACP-safe
